Fix int overflow in decode_coefs()
BUG=b/69205191

(cherry picked from commit 85f2a5ae4c15de5dd530766eb3933b9de976d9cf)

+++ av1/decoder/detokenize.c
@@ -210,7 +210,7 @@ static int decode_coefs(MACROBLOCKD *xd, PLANE_TYPE type, tran_low_t *dqcoeff,
             AOM_QM_BITS;
 #endif
 #if !CONFIG_DAALA_TX
-    v = (val * dqv) >> dq_shift;
+    v = (int)(((int64_t)val * dqv) >> dq_shift);
 #else
     v = val * dqv;
 #endif
